bug(study browser): identify patient by institution and birth date
We noticed that because clinics do not have truly unique user id's, some clinics would see different patients in their related study browser. I've modified the sorting logic to check that a patient is truly a match by including the `institutionName` and `birthDate` as sorting criteria. To do this, I had to update the qido search to include both aforementioned values. This may require an edit in the `default.js` config file stored on ECS. further testing on staging is required.