feat: download study button
This PR adds a button to download the current study. To achieve this, we use the orthanc SHA-1 hashing formula which creates a UUID from the `studyInstanceUID| patientId` and inserts a dash every 8 characters. After that, we create a download action anduse the notification service to display if there were any errors.